+++ payouts/scheme.go
@@ -7,5 +7,5 @@ import (
 )
 
 type Scheme interface {
-	UpdateMinerBalances(poolID string, remainingReward float32, confirmed persistence.Found) (time.Time, error)
+	UpdateMinerBalances(poolID string, remainingReward float64, confirmed persistence.Found) (time.Time, error)
 }
